completed_category

+++ categories/categoriesController.js
@@ -34,6 +34,9 @@ class CategoriesController {
         const {id} = req.params
 
         try {
+            if (isNaN(id))
+                throw new Error('Id is not a number')
+
             const delCategory = await categoriesService.deleteOne(id)
             res.status(200).json(delCategory)
         } catch (e) {
@@ -46,8 +49,31 @@ class CategoriesController {
         const { title, url } = req.body
 
         try {
+            if (isNaN(id))
+                throw new Error('Id is not a number')
+
+            if (title.length < 3 || title.length > 15)
+                throw new Error('Число символов меньше 3 или больше 15!')
+
+            if (!/^(http|https):\/\/([\w-]+(\.[\w-]+)+)(\/[\w-./?%&=]*)?$/.test(url))
+                throw new Error('Ссылка не валидна!')
+
             const upCategory = await categoriesService.update(id, { title, url })
             res.status(200).json(upCategory)
+        } catch (e) {
+            res.status(500).json({message: e.message})
+        }
+    }
+
+    async getOne(req, res) {
+        const { id } = req.params
+
+        try {
+            if (isNaN(id))
+                throw new Error('Id is not a number')
+
+            const category = await categoriesService.getOne(id)
+            res.status(200).json(category)
         } catch (e) {
             res.status(500).json({message: e.message})
         }
